feat: adds `gap_buffer.zig` for `TextInput`
i'm not sure of the best place for `gap_buffer.zig` to
live. definitely don't mind putting it in this repository, but maybe
it might make sense as its own thing, since it's pretty
general-purpose? your thoughts would be very much appreciated. also
tweaked `TextInput` to use a `GapBuffer(u8)` to manage its text. it
_is_ maybe possible for temporarily silly things to happen if a user
intentionally combines two previously-separate graphemes into a single
one somehow? this is because in order to avoid unnecessary O(n)
copies, i assume that the gap is always well-aligned to grapheme
boundaries. i haven't had a chance to take this for a spin yet, since
my program maybe wants the "Multi" version of GapBuffer too... the
gap_buffer.zig file passes all of the tests (which are adapted from
the `std.ArrayList` tests), so it ought to "just work", but thought
you should know.