mouse: always stop handling if event is consumed
The MouseHandler used to only consume the mouse events if there was a
last_handler widget defined AND whoever is using the API set
ctx.consume_event to true.

This change makes it so the event is consumed when ctx.consume_event is
set to true, even if there is no last_handler widget defined.