mouse: set maybe_last_handler when consuming event
Follow-up to my previous patch; this fixes an issue where the mouse_leave
event was never sent because maybe_last_handler was only set when there
was already a maybe_last_handler widget set.

The change here makes it so that every time an event is consumed libvaxis
keeps track of which widget did that through the maybe_last_handler
variable.

In other words: we update maybe_last_handler every time an event is
explcitly consumed.