vxfw: Add ScrollView widget
The widget will render its children in a container that can be scrolled
both horizontally and vertically. The widget itself does not render any
scroll bars or other indicators of current scroll position.

Since this view is heavily based on the ListView widget it inherits the
same `cursor` functionality to show the current position of a selected
widget.

Known Issues
============

1. The view currently does not enforce a maximum width on the content to
   be able to correctly figure out whether the content can still be
   scrolled horizontally. This will cause the widget to draw beyond its
   boundaries horizontally.
2. When the last widget rendered is taller than a single row the whole
   widget will be drawn. This will cause the widget to draw beyond its
   boundaries vertically.